better naming. cleaned up dependencies in cmakelists
