+++ rmcl_ros/include/rmcl_ros/correction/sensors/MICPO1DnSensorCUDA.hpp
@@ -1,5 +1,5 @@
-#ifndef RMCL_MICPO1DN_SENSOR_HPP
-#define RMCL_MICPO1DN_SENSOR_HPP
+#ifndef RMCL_MICPO1DN_SENSOR_GPU_HPP
+#define RMCL_MICPO1DN_SENSOR_GPU_HPP
 
 #include <rclcpp/rclcpp.hpp>
 #include <rmcl_ros/correction/MICPSensor.hpp>
@@ -23,41 +23,35 @@
 
 #include <rmagine/math/statistics.h>
 #include <rmagine/math/linalg.h>
+#include <rmagine/types/MemoryCuda.hpp>
+
 
 #include <geometry_msgs/msg/pose_with_covariance_stamped.hpp>
 
 // Correspondences
 #include <rmcl_ros/correction/Correspondences.hpp>
-#include <rmcl_ros/correction/correspondences/RCCEmbree.hpp>
+#include <rmcl_ros/correction/correspondences/RCCOptix.hpp>
 
 
 #include <mutex>
 #include <thread>
 
 
-
 namespace rmcl
 {
 
-
-
-class MICPO1DnSensor
-: public MICPSensor_<rmagine::RAM>
+class MICPO1DnSensorCUDA
+: public MICPSensor_<rmagine::VRAM_CUDA>
 {
 public:
 
-using Base = MICPSensor_<rmagine::RAM>;
+  using Base = MICPSensor_<rmagine::VRAM_CUDA>;
 
-  MICPO1DnSensor(
+  MICPO1DnSensorCUDA(
     rclcpp::Node::SharedPtr nh,
     std::string topic_name);
 
   void topicCB(const rmcl_msgs::msg::O1DnStamped::SharedPtr msg);
-
-  void setMap(rmagine::EmbreeMapPtr map);
-
-  size_t n_inner_ = 10;
-  size_t n_outer_ = 2;
   
 protected:
 
@@ -65,15 +59,15 @@ protected:
 
 private:
 
-  rmagine::EmbreeMapPtr map_;
+    // we need to store this here, temorary
+  rmagine::PointCloud_<rmagine::RAM> dataset_cpu_;
+
   rmagine::O1DnModel sensor_model_;
 
   message_filters::Subscriber<rmcl_msgs::msg::O1DnStamped> data_sub_;
   std::unique_ptr<tf2_ros::MessageFilter<rmcl_msgs::msg::O1DnStamped> > tf_filter_;
 };
 
-
-
 } // namespace rmcl
 
-#endif // RMCL_MICPO1DN_SENSOR_HPP
+#endif // RMCL_MICPO1DN_SENSOR_GPU_HPP
